README.md: Add "Recommendations" for audio issues.
It's possible that users might run into audio issues while using
confbridge and cmp2k. A section has been added to the README file
("Recommendations") that provides a solution to the problem. In this
case, the jitterbuffer can sort out any disruptive audio issues that
might be encountered while in the conference.